pjm_gen_offer_raw/requirements.txt
requests
pandas
snowflake-connector-python
xlrd